\encoding{UTF-8}
\name{epidemics-package}
\alias{epidemics-package}
\alias{epidemics}
\docType{package}
\title{The epidemics package}

\description{

  This package implements forward-time, individual-based, genetic simulations of pathogen populations.
  
  Essential functionalities of the package are presented in dedicated vignettes accessible using 
  \code{vignette("name-below", package="epidemics")}:\cr
  - epidemics-intro: introduction to the package.\cr
  
  Important functions are also summarized below.\cr

   === SIMULATING DATA ===\cr
  

   === DATASETS ===\cr
  

   For more information, visit the epidemics website:\cr
   \url{http://epidemics.r-forge.r-project.org/} .\cr


   To cite epidemics, please use the reference given by
   \code{citation("epidemics")} (or see reference below).
 }
\details{
  \tabular{ll}{
    Package: \tab epidemics\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2011-09-20\cr
    License: \tab GPL (>=2)
  } 
}
\author{
  Thibaut Jombart <t.jombart@imperial.ac.uk>
}
\references{
 
  See epidemics website: \url{http://epidemics.r-forge.r-project.org/}\cr

  %Please post your questions on 'the epidemics forum': epidemics-forum@lists.r-forge.r-project.org

}
\seealso{
  - \code{ape}, \code{phangorn} and \code{seqinr} packages for phylogenetics and DNA data handling\cr
  - \code{adegenet} and \code{pegas} packages for population genetics tools\cr
}

